+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for `oblique` map orientation.
+- Added `Map` `skew_x` and `skew_y` fields parsed from TMX `skewx`/`skewy`.
+
 ### Changed
 - Switched from `xml-rs` to `quick-xml` to speed up XML parsing.